backend/requirements.txt
# SlopGuard Backend — Python dependencies
# Install: pip install -r requirements.txt

# API framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Supabase client
supabase>=2.7.0

# ML / detection (same as before)
transformers>=4.44.0
torch>=2.3.0
torchvision>=0.18.0
pillow>=10.4.0
numpy>=1.26.0

# Utilities
python-multipart>=0.0.9    # required for file uploads in FastAPI
python-dotenv>=1.0.0       # load .env file
pydantic>=2.8.0            # request/response models
httpx>=0.27.0              # async HTTP (used by supabase client)
